In the middle of some refactoring to reduce shared code between elasticc2 and fastdb_dev.  test_elasticc2_alert_cycle passes
