+++ logs/.placeholder
@@ -1,0 +1,1 @@
+This is here so that the logs directory will be created on git checkout.
